+++ config/description.exs
@@ -662,6 +662,11 @@ config :pleroma, :config_description, [
         description: "Require users to confirm their emails before signing in"
       },
       %{
+        key: :account_approval_required,
+        type: :boolean,
+        description: "Require users to be manually approved by an admin before signing in"
+      },
+      %{
         key: :federating,
         type: :boolean,
         description: "Enable federation with other instances"
@@ -872,6 +877,14 @@ config :pleroma, :config_description, [
         description: "An account field value maximum length. Default: 2048.",
         suggestions: [
           2048
+        ]
+      },
+      %{
+        key: :registration_reason_length,
+        type: :integer,
+        description: "Maximum registration reason length. Default: 500.",
+        suggestions: [
+          500
         ]
       },
       %{