+++ lib/pleroma/user/query.ex
@@ -42,6 +42,7 @@ defmodule Pleroma.User.Query do
             external: boolean(),
             active: boolean(),
             deactivated: boolean(),
+            need_approval: boolean(),
             is_admin: boolean(),
             is_moderator: boolean(),
             super_users: boolean(),
@@ -144,6 +145,10 @@ defmodule Pleroma.User.Query do
   defp compose_query({:deactivated, true}, query) do
     where(query, [u], u.deactivated == ^true)
     |> where([u], not is_nil(u.nickname))
+  end
+
+  defp compose_query({:need_approval, _}, query) do
+    where(query, [u], u.approval_pending)
   end
 
   defp compose_query({:followers, %User{id: id}}, query) do